refactor: sticky pagination component
